+++ src/core/client.ts
@@ -84,7 +84,7 @@ export default class Client {
   }
 
   resetContext(context: Record<string, unknown>): Client {
-    this.logger.warn('Deprecation warning: `Honeybadger.resetContext()` has been deprecated; please use `Honeybadger.clearContext()` instead.')
+    this.logger.warn('Deprecation warning: `Honeybadger.resetContext()` has been deprecated; please use `Honeybadger.clear()` instead.')
     if (typeof context === 'object') {
       this.__context = merge({}, context)
     } else {
@@ -93,7 +93,7 @@ export default class Client {
     return this
   }
 
-  clearContext(): Client {
+  clear(): Client {
     this.__context = {}
     this.__breadcrumbs = []
     return this