+++ src/server/middleware.ts
@@ -16,7 +16,7 @@ function fullUrl(req) {
 }
 
 function requestHandler(req, res, next) {
-  this.clearContext()
+  this.clear()
   const dom = domain.create()
   dom.on('error', next)
   dom.run(next)
@@ -46,7 +46,7 @@ function lambdaHandler(handler) {
     dom.on('error', function(err) {
       hb.notify(err, {
         afterNotify: function() {
-          hb.clearContext()
+          hb.clear()
           callback(err)
         }
       })
@@ -55,11 +55,11 @@ function lambdaHandler(handler) {
     dom.run(function() {
       process.nextTick(function() {
         Promise.resolve(handler.apply(this, args))
-          .then(() => { hb.clearContext() })
+          .then(() => { hb.clear() })
           .catch(function(err) {
             hb.notify(err, {
               afterNotify: function() {
-                hb.clearContext()
+                hb.clear()
                 callback(err)
               }
             })